requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.8
drf-spectacular>=0.26.5
djangorestframework-jwt>=1.11.0
djangorestframework-simplejwt>=5.3.0